pyproject.toml
[project]
name = "stealthmole-for-cisco-xdr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "markdownify>=1.1.0",
    "marshmallow>=3.26.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]
